Improve listing services and pass listings to frontend

+++ src/main/java/com/example/homeXchangeManager/controllers/ListingController.java
@@ -5,28 +5,21 @@ import com.example.homeXchangeManager.models.Listing;
 import com.example.homeXchangeManager.models.User;
 import com.example.homeXchangeManager.repositories.ListingRepository;
 import com.example.homeXchangeManager.repositories.UserRepository;
-import com.example.homeXchangeManager.security.CustomUserDetailsService;
 import com.example.homeXchangeManager.service.impl.UserServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.CachingUserDetailsService;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
-import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
-import java.util.Optional;
 
 @Controller
 public class ListingController {
@@ -89,16 +82,16 @@ public class ListingController {
         return "redirect:/listing";
     }
 
-    @PostMapping("/listing/delete/{id}")
-    public String deleteListing(@PathVariable("id") int listingId) {
-        Listing listing = listingRepository.findListingByListingId(listingId);
-        if (listing != null) {
-            listingRepository.deleteListingByListingId(listingId);
-            logger.debug(String.format("Listing with id: %s has been successfully deleted.", listing.getListingId()));
-            return "redirect:/home_page";
-        } else {
-            // add error pages
-            return "error/404";
-        }
-    }
+
+     @PostMapping("/listing/delete/{id}") public String deleteListing(@PathVariable("id") long listingId) {
+     Listing listing = listingRepository.findListingByListingId(listingId);
+     if (listing != null) {
+     listingRepository.deleteListingByListingId(listingId);
+     logger.debug(String.format("Listing with id: %s has been successfully deleted.", listing.getListingId()));
+     return "redirect:/home_page";
+     } else {
+     // add error pages
+     return "error/404";
+     }
+     }
 }
